Moved damage preset check to function
Now the code to check which damage value to assign to a generic creature is handled by the ChangeGenericMeleeAttack function; it just needs to be passed both the custom damage value and the default damage value now.
Should make sure to include a check if the dictionary key is correct though; have creatures freezing that are missing an entry.