Updated for Nautilus and added Sea Dragon projectile options
Updated the handlers to use the newest version of SMHelper (Nautilus).
Also found how to access the damage values of the projectiles of the SeaDragon; ammoPrefab, located under attackTypes in the RangedAttackLastTarget component of the creature. From there I just needed to navigate to the prefab itself and change its own damage values.