Finished assigning damage values, excluding crashfish and gasopod
All components and damage values, including their default damage values, have been found, coded for and assigned in the dictionary.
This excludes the Crashfish and Gasopod for now; the former because it's private and I need 'Publicizer' to access it, and the latter because it only has a reference to the prefab and not the prefab itself, so can't edit it the usual way.
Still need to clean up the code and create streamlined functions for the different damage assignments, to take into account the damage preset.