ChangeUniqueAttack method functioning
Created a new method, much the same as ChangeGenericMeleeAttack, that takes a reference of the damage value to be changed, instead of defaulting to a basic MeleeAttack component. Otherwise it works exactly the same.
Tested with a Reaper Leviathan.
As such, I may condense much of the doubled-up code into a separate method, to reduce redundant code.