Added Presets & Aggressive Fauna Techtypes
Added the health presets I'm thinking of using for now.

Added all the aggressive fauna from damage mod to the TechType switch to apply the health changes to all of them; still need to update the dictionary for each of their default health values however.

Also, added the word "Leviathan" to all the variables concerning Leviathans; some had them and some didn't, so this is for consistency's sake. Will likely add it back to the damage mod too, with another update.