Fixed blighter coming before blood crawler for both mods
Blighter was missing from the config values on health mod, so have added that.
In doing so, have noticed in alphabetical order, the blighter was in the wrong spot, so have sorted that on both mods.
Also, as was working on default values in the health mod before fixing this, a few config sliders have had a DefaultValue assigned to them.