Implemented Invulnerable function and simplified preset 8 code
Implemented the invulnerable preset and function, which makes the creature's invincible property true, meaning it ignores all damage calculations dealt towards it.
As the Cuddlefish is a pet, I figured having an option specifically for it to be invulnerable would be beneficial.
Finally, as preset 8, invulnerable, makes a creature immune to damage, health calculations are not needed, so it skips most of them if preset 8 is selected.