Copied bomb prefab has working animation and collision
Managed to figure out how the animation triggers work for the bomb prefab, so now when the player approaches one, it starts to open up and vibrate.
Likewise, have made the collision more appropriate for the small bomb that it is.
Aside from still needing to implement all the other behaviour, the only issue is that before spawning a cope of my prefab in, one will be on top of the original bomb in the display case until I spawn it in, in which case it 'vanishes'.