Can detect player collision in water but not air
Found the 'maincollider' property of the player, which matches the collider I'm detecting when the player runs into the bomb *in the water*.
The trouble is, I'm detecting a different collider, 'CharacterController' when on dry land or in the habitat.
May need to once again try and find the root 'Player' object instead.