Able to detect player collisions
Have found a way to detect if the player is the one colliding, by getting the tag of the parent of the collider component, being "Parent".
My only concern going forward is that this component I'm banking on; this PlayerTriggerAnimation, is used elsewhere, and may cause a false-positive; an explosion.