Able to set a modifier to one decimal place
Able to set the min and max size of the creature, randomly selected and remaining at one decimal place.
We do this by multiplying the, for example, min 0.1 and max 1.5 into 1 and 15, then selecting a random integer between them, say 7.
Then we divide back down by 7 and get the final size modifier of 0.7.