Finalised dictionary for fine-tuning min and max size
Finalised the sizeReference dictionary, which the user can use to customise the minimum and maximum random size of every creature.
This is a counterpart to the three Size Class system, which generalises each of the creatures into three groups, but allows for much quicker and easier customisation of minimum and maximum size range.