Added attempt at factoring in maturity of WPC
Attempted to make sure maturity was factored in, when setting the size of a creature; as a juvenile would be smaller than their max size, they would need to be calculated from there.

However, much like other properties, maturity is not set at runtime, so this will have to be a pending issue.