Found method to check if creature is inside containment
By checking if a parent of the creature has the WaterPark component (the LargeRoomWaterPark component is a derivative of the WaterPark component, so it works in both cases), I can determine at start if a creature is in containment.

This likely means I can perform all checks are Creature.Start, instead of fragmenting them across multiple patches.