Added function to get size class
Added a function to return the size class of a creature; small, medium or large (or none, if its techtype is not in any of the arrays).

Have so far used this function to check if a given creature is a leviathan, and preventing them from being able to breed in containment.