Realised error with using LiveMixin.Awake
Realised that by using LiveMixin.Awake, for some reason, by the time the creature is active in-game, the size changes have been overridden; possibly by Creature.Start?

Also, Reefback and Reefback Juveniles share the same TechType, despite other instances where they don't, so I'm working on a test to change their TechType if need be, for better or for worse.