Created a fix to issues with Set Size Command
Created a fix to the issue with the set size command, specifically by targeting the object or parent object with the creature component, which solved the issue of raycasting child objects with no creature component, such as with the Mesmer.
Unfortunately, this has caused an unknown error to occur when no target is selected, which I will have to fix.