Updated shuttlebug size and plan for fixing issues
Made the blood crawler a bit smaller, as it looks a bit better in the smaller containment when it's not clipping into the ceiling.

Also found that turning back on the creature component for blood crawlers, cave crawlers and seadragons fixes their lack of pathing and animation issues, so will likely be going that route.
Also will be doing that for skyrays, so they drown if someone is mean enough to try to put one in a watertank.