Testing fix to first-time size change
Found that many of the basic fauna start at sizes beyond 1.0 (can be as random as to 3 decimal places), so that doesn't work for many creatures.

Two ideas I have are to either check if a creature has a size beyond 1 decimal place, as none of my code should do this. However, this will have issues with fish growing in the alien containment, and will likely need some loopholes there to account.

The other suggestion is a custom component that saves a boolean between game instances, but I added it to a creature and the component was gone when I reloaded the game, meaning it's currently uncertain whether I can save the state of a boolean between instances of the game.