Removed filter for no techtype and fish school
Because I'm now filtering from the beginning only creature's I've specified in the Pickupable reference table, I now no longer need to filter further for anything that lacks a TechType, or is a fish school, because they're already *not* in the reference table I'm going off of.

This saves on some checks and makes the code nicer.