Added creature info class, list and method
Created a class, with constructor and deconstructor to streamline adding and retrivieving entries, to store all the info each creature needs to properly set them up as an inventory item.

Also created a list to store said info, which currently has only 3 entries so far.

Also created a method to pull from the list and automatically assign the info to the TechTypes provided in the list.