Updated both component check functions
Updated both functions that check for pickupable and waterparkcreature components.

Pickupable now also checks for if a creature is still in containment, and will ensure it has the pickupable component if so.

WaterParkCreature now also does the same in checking if the creature is inside containment, but also ensures that the WaterParkCreatureData is filled in, regardless of whether a creature should or should not be in containment, to make sure there are no massive creatures somehow in containment.