Test to add child prefab from GUID
Tested using GUID to add child prefab to spawn, having a reaper birth peepers.

This was done by retrieving the GUID of peepers, that I found because they already have them in their WPC data, as they live-birth, and retrieved an AssetReferenceGameObject from that GUID and added that to the Reaper's eggOrChildPrefab.

This allowed it to birth peepers.

I have played God.