Tested various methods of getting child prefab - Method 5 works!
Tested 5 different methods of getting an AssetReferenceGameObject that I could add to eggOrChildPrefab to spawn in.

The secret sauce to success was Method 5, getting a Reference using the filepath "WorldEntities/Creatures/SandShark.prefab", then using '.ForceValid()' to make sure the RuntimeKey was valid and passed the check to spawn in.