Used AdultPrefab to have SeaEmperorBaby age to Juvenile
By making the AdultPrefab of a SeaEmperorBaby a reference to the SeaEmperorJuvenile prefab, and making it not able to breed, it will eventually age *into* a juvenile!

This means I can have the baby/juvenile versions of some of the creatures eventually age into their adult forms!

I should be wary of size increases/decreases however and make them relatively the same scale as their baby form, if possible.