child & adult prefabs assigned automatically now
Egg/Child prefabs and Adult prefabs are now being automatically assigned to each creature, the former if their techtype is in the AssetPrefabReference dictionary, and the latter if they're part of a switchcase statement.

Also, just added the prefabs to the dictionary with '.ForceValid()' ensuring they're valid right from the get-go!