Fixed IL code and enabled reaper to death grab
Fixed the IL code, courtesy of the help of the modding discord, and have now succeeded in replacing the null 'dealer' value the Reaper death grab passes to a reference to the reaper itself, meaning I can detect if they're to ignore the reinforced dive suit and properly trigger the death grab if the damage exceeds the player's current health.

Still need to apply these changes to the other creatures that have a death grab; Sea Dragons, and possible Crab Snakes.